Machine_Learning_Module/packagePrediction/requirement.txt
google-cloud-aiplatform
google-cloud-firestore
flask
flask_cors
firebase_admin

